Moved the training functions to the gan file.
